refactor: Create services folder in desktop/electron-app
