+++ ui/desktop/electron-app/src/services/session-manager.js
